Fixing Text rendered behind borderLayer (#46065)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/46065

This change fixes an issue that has been reported by OSS where a Text with both background color and borderWidth is not rendered properly.

The reason is that `RCTParagraphComponentView` uses the `drawRect` method which draws the text in the main view layer, while the parent `RCTViewComponentView` can apply an extraLayer on top of the base layer, drawing on top of the text.

Tis change makes sure that the background color is set in the same base layer that is used to draw the text, while the borderLayer always has a transparent background.

## Changelog
[iOS][Fixed] - make sure that the text is always visible when applying background color and border width

Differential Revision: D61390097