fix[DebuggingOverlayRegistry]: dont use microtasks for legacy stack (#46279)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/46279

# Changelog: [Internal]

We've discovered that highlighting feature is a bit bugged when you use for RNTester: for some reason highlights would persist on the screen, even if the React DevTools panel was unfocused. This wasn't reproducible on Catalyst, though.

In highlighting case, we wrap `.measure()` calll in a `Promise`, in order to later call `Promise.all` to resolve them all together and get a collection of boxes on the screen that we should highlight. I've discovered that callbacks in `.then()` and `.catch()` are never executed, unless a macrotask is scheduled after it. Microtask queue is not drained for some reason.

This migrates the implementation from using Promises and microtasks.

|  | Before | After |
| Highlighting |  https://pxl.cl/5vbZ8 |   https://pxl.cl/5vbxR |
| Tracing updates |  https://pxl.cl/5vbZ6 | {F1836251140} |

Differential Revision: D62032627