Sync Fabric Renderers (#46303)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/46303

This change is in preparation for the branch cut of 0.76. This syncs the renderers between React and React Native

## Changelog
[Internal] - Sync React renderers

## Facebook
This change is required to fix [#42101](https://github.com/facebook/react-native/issues/42101). You can read more about that [here](https://fb.workplace.com/groups/615693552291894/permalink/1806859756508595/)

I'll enable the flag in the next diff

Reviewed By: blakef

Differential Revision: D62100659